Add AWSConfig cog for storing shared AWS configs
AWSConfig is a simple cog used for storing AWS access keys/secrets and
region information which can be retrieved by any cog which has support
added.

Information is stored in the cog's GLOBAL config key, and consists of
aws_access_key_id, aws_secret_access_key, and region. This information
is sufficient for establishing a session with AWS through boto3.

The configuration can be retrieved in any sibling cog by using:

    awsconfig_cog = self.bot.get_cog("AWSConfig")
    aws_config = awsconfig_cog.get_aws_config() if awsconfig_cog else {}

Consumer cogs will be responsible for fallbacks or instructing the end
user to set the AWS configuration using the AWSConfig cog commands if
the config is not set.

If you are attempting to load the configuration during a cog
initialization, there is no guarantee that the AWSConfig cog has already
been loaded. In this case, it is best to use asyncio's wait_for method
to loop until AWSConfig is loaded, or until a timeout has passed.

Example:

    async def _ensure_awsconfig():
        while True:
            if self.bot.get_cog("AWSconfig") is not None:
                return
            asyncio.sleep(0.1)
    await asyncio.wait_for(_ensure_awsconfig(), timeout=5)
    awsconfig_cog = self.bot.get_cog("AWSConfig")