Force AWSConfig cache to load if it is empty
If get_aws_config is called on AWSConfig before the _init_cache future
has finished running, it will reuturn an empty dict. This is a problem,
for example, when the cache is needed during the initialization of other
cogs.

This change forces the cache to load from the database if it is empty
and a get is called on the cache from elsewhere.