Add support in seplib for logging to AWS CloudWatch
Adds a new coroutine to seplib which can be used to attach a new handler
to a python Logger which will send all logs to CloudWatch Logs.

The AWS account is configured through the AWSConfig cog, and the IAM
account for the credentials must have the following policies:

Effect: Allow
Action:
  logs:CreateLogGroup
  logs:CreateLogStream
  logs:DescribeLogStreams
  logs:PutLogEvents
Resource: *

A new log group will be created with the name format
"/discord/red/{BOT_USER_NAME#DISCRIMINATOR}". A new log stream will be
created for each cog utilizing the handler with the name of the Logger.

CloudWatch logging can be attached to any cog by doing something similar
to the following:

    class MyCog(commands.Cog):
         def __init__(self, bot):
            self.bot = bot
            self.logger = logging.getLogger(f"red.MyRepo.mycog")
            self.logger.setLevel(logging.INFO)
            asyncio.ensure_future(set_cloudwatch_handler(cog=self,
bot=self.bot, logger=self.logger)